requirements.txt
plotly.express == 0.4.1
plotly == 5.17.0
pandas == 2.1.1
streamlit == 1.28.2
folium == 0.14.0
numpy == 1.24.4
pil_plus == 0.2
streamlit_folium == 0.13.0
haversine == 2.8.0
inflection==0.5.1

